app/.dev/DevOpsPython/requirements.txt
google-api-python-client==2.39.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
httplib2==0.22.0
PyYAML==6.0.1
